backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
transformers==4.38.1
torch==2.2.0
sentencepiece==0.2.0
accelerate==0.27.2
python-multipart==0.0.9
pydantic==2.6.1
cors==1.0.1
streamlit==1.31.1
httpx==0.24.1 